Updated to use JSX
