+++ packages/snap/package.json
@@ -28,7 +28,7 @@
     "test": "jest"
   },
   "dependencies": {
-    "@metamask/snaps-sdk": "^4.0.0",
+    "@metamask/snaps-sdk": "^4.4.2",
     "buffer": "^6.0.3"
   },
   "devDependencies": {
@@ -38,7 +38,7 @@
     "@metamask/eslint-config-jest": "^12.1.0",
     "@metamask/eslint-config-nodejs": "^12.1.0",
     "@metamask/eslint-config-typescript": "^12.1.0",
-    "@metamask/snaps-cli": "^6.1.0",
+    "@metamask/snaps-cli": "^6.2.0",
     "@metamask/snaps-jest": "^7.0.0",
     "@typescript-eslint/eslint-plugin": "^5.42.1",
     "@typescript-eslint/parser": "^5.42.1",
