+++ packages/snap/snap.manifest.json
@@ -7,7 +7,7 @@
     "url": "https://github.com/Montoya/hello-world-snap.git"
   },
   "source": {
-    "shasum": "dfOSXvkVrH6GVSAmVtw34IWZYkPKcJtJrv3eHfo2cQQ=",
+    "shasum": "jdJ17no1St4VNrSQpbCm/B04sZzIHkuLh56pwUqXb6c=",
     "location": {
       "npm": {
         "filePath": "dist/bundle.js",
